+++ lib/sec_id/fisn.rb
@@ -24,6 +24,7 @@ module SecId
     FULL_NAME = 'Financial Instrument Short Name'
     ID_LENGTH = (3..35)
     EXAMPLE = 'APPLE INC/SH'
+    VALID_CHARS_REGEX = %r{\A[A-Z0-9 /]+\z}
 
     # Regular expression for parsing FISN components.
     # Issuer: 1-15 chars, Description: 1-19 chars, Total: max 35 chars